refactor(buffers): split BufferArena accessors to a sibling file; allow-list TransformerLayer trait
Fix the two ≤500-LoC violations the LoRA MVP introduced:

- spark-runtime/src/buffers.rs (555 -> 393): the LoRA scratch buffers
  pushed it over. Extract the pointer/size accessor getters (a cohesive
  cluster, including the lora_xa/lora_delta/lora_hact getters) into a new
  buffers/accessors.rs as a second impl block (child module reads the
  private fields). Real code split per Atlas idiom.

- spark-model/src/layer/transformer_layer.rs (504 -> 502): the as_any_mut
  downcast hook (needed by the LoRA install walk) tipped this trait over.
  It is an atomic interface — ~20 arg-heavy methods whose SSM-prefill
  defaults are one-line bail/Ok no-ops, so body-extraction is net-neutral
  (same situation as the already-allow-listed traits/model.rs). Trim the
  as_any_mut doc and allow-list it with rationale; a supertrait split is
  tracked as follow-up.
