+++ crates/spark-model/src/layer/transformer_layer.rs
@@ -12,10 +12,8 @@ use super::{BatchedAttnMetadata, ForwardContext, GdnPrefillBuffers, LayerState};
 mod default_loops;
 
 pub trait TransformerLayer: Send + Sync {
-    /// `&mut dyn Any` downcast hook for post-construction weight overlays
-    /// (e.g. the LoRA install walk downcasting to `Qwen3AttentionLayer`).
-    /// Default `None`; concrete layers that support downcast-based installs
-    /// override with `Some(self)`.
+    /// `&mut dyn Any` downcast hook for post-construction weight overlays (e.g.
+    /// the LoRA install walk). Default `None`; overlay-capable layers override.
     fn as_any_mut(&mut self) -> Option<&mut dyn std::any::Any> {
         None
     }
